comments for clarity on how the C wrapper works. Name change to the Go file implementation
